implemented connecting and disconnecting. Connect still need work
